requirements.txt
fastapi==0.109.2
langchain== 1.1.2
langchain-openai==1.1.0
langchain-pinecone==0.2.13
pinecone==7.3.0
boto3==1.34.131
twilio==9.5.1
python-dotenv==1.0.1
uvicorn==0.30.3
python-multipart==0.0.21
openai==2.9.0
requests==2.32.5